+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.0.1
+
+- Fix markdown rendering issue
+- Fix highlight code blocks rendering issue
+- Add copy button to code blocks
+- Add language label to code blocks
+
 ## v1.0.0
 
 - Initial release